requirements.txt
nicegui
requests
python-dotenv
pydantic
uvicorn
